Add- AMP binarysearch, Benchmark script, DataType.hpp
