+++ bench/benchmark.py
@@ -1,0 +1,125 @@
+
+print("###Copy the clBolt.Bench.Benchmark.exe to a separate folder along with all python scripts and run this batch script")  
+
+print("### This will benchmark the Routines for DATA_TYPE which is a MACRO in the Benchmark.cpp file. Set it to the Appropriate\n\
+\tdata type for which you wan to Benchmark. ie unsigned int, float, double etc.\n\
+\tAnd build Benchmark project to get the Required clBolt.Bench.Benchmark.exe")
+print("### This will create the some comparison graphs ")
+print("### BUT Before Running this script make sure your system must have PYTHON installed ( python-2.7.3.msi )")
+print("### Installed all these required useful package as well - \n\
+          \t\t\tnumpy-1.8.0-win32-superpack-python2.7.exe\n\
+		  \t\tscipy-0.13.1-win32-superpack-python2.7.exe\n\
+ 		  \t\twxPython2.8-win32-unicode-2.8.12.1-py27.exe\n\
+ 		  \t\tmatplotlib-1.3.1.win32-py2.7.exe\n\
+ 		  \t\tpython-tesseract_0.8-1.5.win32-py2.7.exe\n\
+ 		  \t\tpython-dateutil-2.2.win32-py2.7.exe\n\
+ 		  \t\tpyparsing-2.0.1.win32-py2.7.exe\n\
+ 		  \t\tsix-1.4.1.win32-py2.7.exe")
+ 
+print("### Default value of Iterator = 50, to change the no. of iteration, please go to line no. 325 of file commonPeformance.py")
+print("### In command-line-arguments -l a1-a1:x2  --> means for each iteration double the length of buffer w.r.t. previous length.\n\
+					\t-l a1-a2:a3  --> means for each iteration increase the buffer size by a3.")
+
+import os
+import sys
+import argparse
+import glob
+
+
+parser = argparse.ArgumentParser()
+parser.add_argument("-r","--routine", help="routine name")
+parser.add_argument("-l","--lenghtOFBuffer", help="Length of buffer")
+parser.add_argument("-i","--iteration", help="no. of iteration per routine")
+args = parser.parse_args()
+if (args.iteration and args.iteration and args.routine):
+
+	txtpdf_files = glob.glob('*.txt*')
+	print txtpdf_files
+	deleteFile = raw_input("Would you like to delete all .txt and .pdf files?(yes/no) ")
+	if deleteFile == "yes":
+		# need to loop through txtpdf_files and use os.remove
+		length = len(txtpdf_files);
+		while length > 0:
+			os.remove(txtpdf_files[length-1])
+			length = length -1;
+	elif deleteFile == "no":
+		print "\nCAVEAT: If .txt file of same routine exist, it will create new .txt file with some random number appended to it\
+		\n\tWhich lead to wrong pdf file creation"
+			
+	#python commonPerformance.py --label reduce_bolt_device --library=BOLT --routine=reduce --memory device -l 1024-33554432:x2 --tablefile reduce_bolt_device.txt
+	string = " --label "\
+	+sys.argv[2]+"_bolt_device --library=BOLT --routine="\
+	+sys.argv[2]+" --memory device -l "\
+	+sys.argv[4]+" -i "\
+	+sys.argv[6]+" --tablefile "\
+	+sys.argv[2]+"_bolt_device.txt";
+	os.system('python commonPerformance.py%s'%string)
+	
+	string = " --label "\
+	+sys.argv[2]+"_bolt_host --library=BOLT --routine="\
+	+sys.argv[2]+" --memory host -l "\
+	+sys.argv[4]+" -i "\
+	+sys.argv[6]+" --tablefile "\
+	+sys.argv[2]+"_bolt_host.txt";
+	os.system('python commonPerformance.py%s'%string)
+	
+	string = " --label "\
+	+sys.argv[2]+"_tbb_host --library=TBB --routine="\
+	+sys.argv[2]+" --memory host -l "\
+	+sys.argv[4]+" -i "\
+	+sys.argv[6]+" --tablefile "\
+	+sys.argv[2]+"_tbb_host.txt";
+	os.system('python commonPerformance.py%s'%string)
+	
+	string = " --label "\
+	+sys.argv[2]+"_stl_host --library=STL --routine="\
+	+sys.argv[2]+" --memory host -l "\
+	+sys.argv[4]+" -i "\
+	+sys.argv[6]+" --tablefile "\
+	+sys.argv[2]+"_stl_host.txt";
+	os.system('python commonPerformance.py%s'%string)
+	
+	#python plotPerformance.py --y_axis_label "MKeys/sec" --title "sort Performance" --x_axis_scale log2 -d sort_bolt_device.txt -d sort_bolt_host.txt -d sort_tbb_host.txt -d sort_stl_host.txt --outputfile sortPerf_boltDevice_boltHost_tbbHost_stl_Holst.pdf	
+	string = " --y_axis_label \"MKeys/sec\" --title \""\
+	+sys.argv[2]+" Performance\" --x_axis_scale log2 -d "\
+	+sys.argv[2]+"_bolt_device.txt -d "\
+	+sys.argv[2]+"_bolt_host.txt -d "\
+	+sys.argv[2]+"_tbb_host.txt -d "\
+	+sys.argv[2]+"_stl_host.txt --outputfile "\
+	+sys.argv[2]+"Perf_boltDevice_boltHost_tbbHost_stl_Host.pdf";
+	os.system('python plotPerformance.py%s'%string)
+	
+	print("generate the graph for Bolt_With_DeviceVector VS Tbb_With_HostVector")
+	string = " --y_axis_label \"MKeys/sec\" --title \""\
+	+sys.argv[2]+" Performance\" --x_axis_scale log2 -d "\
+	+sys.argv[2]+"_bolt_device.txt -d "\
+	+sys.argv[2]+"_tbb_host.txt --outputfile "\
+	+sys.argv[2]+"Perf_boltDevice_tbbHost.pdf";
+	os.system('python plotPerformance.py%s'%string)
+	
+	print("generate the individual graph for Bolt_With_DeviceVector")
+	string = " --y_axis_label \"MKeys/sec\" --title \""\
+	+sys.argv[2]+" Performance\" --x_axis_scale log2 -d "\
+	+sys.argv[2]+"_bolt_device.txt --outputfile "\
+	+sys.argv[2]+"Perf_boltDevice.pdf";
+	os.system('python plotPerformance.py%s'%string)
+else:
+	print "Too few arguments: Use -h or --help as input argument to HELP you"
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
