+++ test/amp/BinarySearch/targetver.h
@@ -1,0 +1,25 @@
+/***************************************************************************                                                                                     
+*   Copyright 2012 - 2013 Advanced Micro Devices, Inc.                                     
+*                                                                                    
+*   Licensed under the Apache License, Version 2.0 (the "License");   
+*   you may not use this file except in compliance with the License.                 
+*   You may obtain a copy of the License at                                          
+*                                                                                    
+*       http://www.apache.org/licenses/LICENSE-2.0                      
+*                                                                                    
+*   Unless required by applicable law or agreed to in writing, software              
+*   distributed under the License is distributed on an "AS IS" BASIS,              
+*   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.         
+*   See the License for the specific language governing permissions and              
+*   limitations under the License.                                                   
+
+***************************************************************************/                                                                                     
+
+#pragma once
+
+// Including SDKDDKVer.h defines the highest available Windows platform.
+
+// To build your application for a previous Windows platform, include WinSDKVer.h, and,
+//  before including SDKDDKVer.h, set the _WIN32_WINNT macro to the platform you want to support.
+
+#include <SDKDDKVer.h>
