+++ test/amp/CMakeLists.txt
@@ -63,3 +63,4 @@ add_subdirectory( SortTest )
 add_subdirectory( TransformTest )   
 add_subdirectory( TransformReduceTest )
 add_subdirectory( device_vector )
+add_subdirectory( BinarySearch )
